chore(phase-3): split sqlite-helpers tests, wire 2 orphan audits into CI
- Phase 2: catalog audit complete. 164 entries triaged (DELETE-CODE/ADD-TEST/
  KEEP-IRREDUCIBLE/KEEP-INFRASTRUCTURE). 122 entries are genuine equivalents
  (crypto round-trips, boundary conditions, type-narrowing with downstream
  filter, SQL template invariants); 22 are vitest worker-cache static-mutation
  traps documented in project-static-mutants-empirical.md. 3 entries
  (admin-ui.ts L24/L202/L206) are dead-in-tests fallbacks classified as
  ADD-TEST but deferred — proper fix needs production-config-absent fixture.
- Phase 3B: extracted pure-helper tests (resolveMetaString, resolveSqliteStatus,
  snapshotField, buildRevisionParams, SQL constants) into
  sqlite-helpers-utils.test.ts. sqlite-helpers-unit.test.ts is now 1635 lines
  of integration flow tests vs the prior 1780 mixed-purpose lines.
- Phase 3C: wired audit:integration-honesty and audit:integration-secrets into
  the CI parallel-audits batch. audit:bundle stays (referenced in
  EVALUATION.md, no-op without dist artifact). audit:discovery and
  audit:followup were already in lefthook (plan was based on stale view).
- Phase 3D: already implemented — .stryker-tmp/sandbox-* is in sweep-tmp
  lefthook step (plan listed it for stale info).
